requirements.txt
streamlit>=1.45.0
openai>=1.109.1
qdrant-client>=1.17.1
pypdf>=6.9.1
python-dotenv>=1.2.2
tiktoken>=0.12.0
